Add option to skip storing state of earlier blocks during sync
The new configuration option "-omniskipstoringstate" allows to set a block until which state is only stored to disk every 5000 blocks.

If using this option with a more recent block, this can speed up the initial synchronization/parsing of the Omni state. However, if there is an expected or unexpected shutdown before reaching this point, the most frequent state will be lost and has to be reparsed.